+++ library/src/main/java/com/getstream/sdk/chat/view/MessageInputView.kt
@@ -2,6 +2,7 @@ package com.getstream.sdk.chat.view
 
 import android.app.Activity
 import android.content.Context
+import android.net.Uri
 import android.os.Bundle
 import android.text.Editable
 import android.text.TextUtils
@@ -20,6 +21,7 @@ import androidx.recyclerview.widget.ConcatAdapter
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import com.getstream.sdk.chat.CaptureMediaContract
+import com.getstream.sdk.chat.DocumentTreeAccessContract
 import com.getstream.sdk.chat.adapter.CommandsAdapter
 import com.getstream.sdk.chat.adapter.MentionsAdapter
 import com.getstream.sdk.chat.databinding.StreamViewMessageInputBinding
@@ -28,10 +30,10 @@ import com.getstream.sdk.chat.model.AttachmentMetaData
 import com.getstream.sdk.chat.model.ModelType
 import com.getstream.sdk.chat.utils.InputMode
 import com.getstream.sdk.chat.utils.MessageInputController
-import com.getstream.sdk.chat.utils.PermissionChecker
 import com.getstream.sdk.chat.utils.StringUtility
 import com.getstream.sdk.chat.utils.TextViewUtils
 import com.getstream.sdk.chat.utils.Utils
+import com.getstream.sdk.chat.view.common.visible
 import com.getstream.sdk.chat.whenFalse
 import com.getstream.sdk.chat.whenTrue
 import io.getstream.chat.android.client.models.Attachment
@@ -50,11 +52,6 @@ class MessageInputView(context: Context, attrs: AttributeSet?) : RelativeLayout(
      * Styling class for the MessageInput
      */
     private val style: MessageInputStyle = MessageInputStyle(context, attrs)
-
-    /**
-     * Permission Request listener
-     */
-    private var permissionRequestListener: PermissionRequestListener? = null
 
     private var isKeyboardEventListenerInitialized = false
 
@@ -94,6 +91,17 @@ class MessageInputView(context: Context, attrs: AttributeSet?) : RelativeLayout(
             ?.registerForActivityResult(CaptureMediaContract()) { file: File? ->
                 file?.let { messageInputController.onFileCaptured(it) }
             }
+
+    private val documentTreeAccessContract: ActivityResultLauncher<Unit>? =
+        (context as? ComponentActivity)
+            ?.registerForActivityResult(DocumentTreeAccessContract()) { uri: Uri? ->
+                messageInputController.onClickOpenSelectView(
+                    null,
+                    false,
+                    uri
+                )
+            }
+
     private val commandsAdapter =
         CommandsAdapter(style) { messageInputController.onCommandSelected(it) }
     private val mentionsAdapter = MentionsAdapter(style) {
@@ -115,8 +123,7 @@ class MessageInputView(context: Context, attrs: AttributeSet?) : RelativeLayout(
     private fun applyStyle() {
         ActivityResultContracts.GetContent()
         // Attachment Button
-        binding.ivOpenAttach.visibility =
-            if (style.isShowAttachmentButton) View.VISIBLE else View.GONE
+        binding.ivOpenAttach.visible(style.isShowAttachmentButton)
         binding.ivOpenAttach.setImageDrawable(style.getAttachmentButtonIcon(false))
         binding.ivOpenAttach.layoutParams.width = style.attachmentButtonWidth
         binding.ivOpenAttach.layoutParams.height = style.attachmentButtonHeight
@@ -129,6 +136,7 @@ class MessageInputView(context: Context, attrs: AttributeSet?) : RelativeLayout(
         binding.llComposer.background = style.inputBackground
         // Input Text
         style.inputText.apply(binding.messageTextInput)
+        binding.messageTextInput.hint = style.getInputHint()
         style.inputBackgroundText.apply(binding.tvTitle)
         style.inputBackgroundText.apply(binding.tvCommand)
         style.inputBackgroundText.apply(binding.tvUploadPhotoVideo)
@@ -140,9 +148,6 @@ class MessageInputView(context: Context, attrs: AttributeSet?) : RelativeLayout(
         binding.sendButton.setOnClickListener { onSendMessage() }
         binding.ivOpenAttach.setOnClickListener {
             messageInputController.onClickOpenBackGroundView(MessageInputType.ADD_FILE)
-            if (!PermissionChecker.isGrantedCameraPermissions(context) &&
-                permissionRequestListener != null && !style.passedPermissionCheck()
-            ) permissionRequestListener!!.openPermissionRequest()
         }
     }
 
@@ -176,9 +181,8 @@ class MessageInputView(context: Context, attrs: AttributeSet?) : RelativeLayout(
 
     private fun configSendButtonEnableState() {
         val attachments = messageInputController.getSelectedAttachments()
-        val hasAttachment = attachments.isNotEmpty()
         val notEmptyMessage =
-            !StringUtility.isEmptyTextMessage(messageText) || !messageInputController.isUploadingFile && hasAttachment
+            !StringUtility.isEmptyTextMessage(messageText) || attachments.isNotEmpty()
         binding.activeMessageSend = notEmptyMessage
     }
 
@@ -198,10 +202,7 @@ class MessageInputView(context: Context, attrs: AttributeSet?) : RelativeLayout(
         }
         binding.llCamera.setOnClickListener { messageInputController.onCameraClick() }
         binding.llFile.setOnClickListener {
-            messageInputController.onClickOpenSelectView(
-                null,
-                false
-            )
+            documentTreeAccessContract?.launch(Unit)
         }
     }
 
@@ -339,10 +340,6 @@ class MessageInputView(context: Context, attrs: AttributeSet?) : RelativeLayout(
     interface TypeListener {
         fun onKeystroke()
         fun onStopTyping()
-    }
-
-    interface PermissionRequestListener {
-        fun openPermissionRequest()
     }
 
     interface MessageSendHandler {