+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # To be released:
 - Implement Giphy actions handler
 
+# Sep 4th, 2020 - 4.2.11-beta-11
+
+- Fix uploading files and capturing images on Android >= 10
+- Fix `AvatarView`: Render lastActiveUsers avatars when channel image is not present
+
 # Aug 28th, 2020 - 4.2.11-beta-9
 
 - Update event structure